Fix broken tests due to changes.
import_akomantoso and import_json tests:
* Changes in popit-resolver mean we need to do more set up here.
* override_settings does not override in setUpClass (and it's not really
the setting that is being used here, only as a variable lookup), so pass
in a value directly.
* Maintain test separation by removing objects created in setUpClass
  during tearDownClass.
* Make sure to clean the search index at the start of any test using it.

populate_speakers_command_tests: override the setting correctly.

search_tests: Again, clear the index, and use the ID directly (a change
due to using elasticsearch even in tests).