Display new speaker after an error/ when editing.
If a form is submitted with a speaker to be created but there is
another error somewhere on the form then a little dance needs to
be performed in order to get the new speaker displayed correctly
in the new bound form.

Also check when editing a speech, we display the speaker.