Check that speech text is preserved on removing/adding a speaker.
A 'speech' with no speaker is represented in the Akoma Ntoso as
a <narrative> element, which can't contain <p>, whereas a speech
with a speaker is made up of paragraphs.

When someone changes a speech from having a speaker to not or vice
versa, we therefore have to add/remove <p> elements.

This test is to make sure that if we add a speech with a speaker,
remove the speaker, add them back again, and then remove them again,
that we always have the same two variants of the text.