Move scraper base classes to SayIt.
This module used to live in sayit.mysociety.org, but as it is of
more general use we're shifting it to sayit. It shouldn't be considered
to be part of a stable API at sayit though, so I've added a note to
that effect in the docstring of the module.

These scraper helper classes need a couple of extra python packages.
As these aren't necessary for every user of SayIt, we've put them
in a separate requirements file

requirements-scraping.txt.