South and popit-django are not required.
Without South present, syncdb will create up-to-date tables, which is
okay for a fresh install. A user of the SayIt app who uses South will
also have to install popit-django for the migrations to fully run
(unless they get up to date without South and then fake the migration
run).