Be prepared for custom user models.
Switch from using django.contrib.auth.models.User to
django.contrib.auth.get_user_model to make things easier when someone
wants to use sayit with a custom user model.

As this is all in the test helper it probably doesn't actually matter
but it seems like good practice to do this, so we may as well start.