Make speech css class dependent on speech type.
Instead of doing an if based on whether or not there is a speaker,
use the type of a speech to name the css class. This should preserve
the current classes based on 'narrative' and 'speech' for speeches
with or without speakers, but will allow speeches which have other
types to have sensible classes for the future.