Add a date picker for speech and section forms.
Bootstrap has been removed, but we were still trying to use the
bootstrap date and time pickers. Replace the date picker with
a foundation one. For the moment we'll just put the time picker
back to a text box with a placeholder.

I can't see autoclose mentioned in the fdatepicker docs, so I've
taken that out - the new datepicker seems to close when a date is
selected anyway.

Fixes #366.