Don't set a minimum end date based on start date.
With this as it was, if the start date were erroneously set too far
in the future, and then brought back, the picker for the end date
was stuck with only accepting dates after the original start date.

We should have a proper think about how to make the javascript keep
the dates and times consistent.