Add OpenGraph data and tests.
Add Opengraph data to the base template, so it appears on every page.
In particular, tailor it to
* instance home pages
* speaker detail pages
* speech detail pages
* section detail pages

Allow testing of OG keys against regular expressions or strings.

A dictionary of OG keys to tests can be passed in to a utility
function for testing the OpenGraph data at a URL. The keys are
OpenGraph keys without the initial 'og:', the values are either
strings of the expected output or regular expressions which must
be found in the expected output by searching.

Additionally any OG key which should contain a URL is checked to
make sure the URL is absolute.