Only import persons.
Currently, persons are the only objects from popolo that sayit
uses, and they are imported as speeches.Speaker objects, which
are popolo.Person objects which are subdomain instance aware.

Rather than making instance aware versions of the other Popolo
object types, for the moment we're going to import only the
Persons.