Break popolo importer out of management command.
Break the Popolo importer out of the management command so we
can use it from the web frontend too.

The importer can now be created with an instance to import
stuff into, and this can be passed in from the management command.

It's also been updated to work with a single file or unpaginated output
like

http://scrapers-ruby.herokuapp.com/organizations

Some extra python packages are needed in order to download from Popit
using requests.