Print out stats in command line AN import.
import_document() has not been returning a section for some time.
It's possible there isn't one "top level" section in an AN file (if
there's no docTitle, for example), so output the stats that the web
interface to AN import is currently using.

Thanks to Jaroslav Semančík for reporting. Fixes #460.